removing GoGaf21
